refactor: post manager 컴포넌트 분리
