+++ src/pages/post-manager/ui/index.ts
@@ -1,0 +1,1 @@
+export * from "./post-manager"
